Add engine test cases
Starting with the main logics, the merge functionality, the test-cases
presented in the requirements.md are used as test parameters. Maybe
other tests need to be added to cover edge cases, but these seem to be
a very reasonable start. It does not look like other cases are actually
needed as the mechanic is quite simple.

All tests are still failing as the minimal merge implementation returns
the given, unmerged state.

Curiously, there was an issue with running the black formatter due to
naming the files with the types "types.py" apparently this can be an
issue. Therefore, the file is named "typings.py" instead.
See: https://stackoverflow.com/questions/44463879/could-not-import-runpy-module