Implement merge up and down
Quite similar to merge right, we can solve the upward and downward
direction merge by combining flipping and rotating the state before
and after the merge.

Since there is no automated typecheck yet (and probably never will be
due to the scope of this assignment), we keep the UnsupportedDirection
exception, just in case. Add a test for this case.