Add a generator for the initial state
The initial state can be generated by supplying the size of the play
area (to simplify only square for this exercise), a list of possible
values that will be chosen from randomly by the supplied choice
function.

There is a constraint in place that the generated play area cannot be
empty.