Add an evolve function to perform one game step
Evolve handles all necessary game logic to perform an individual step in
the game from a chosen direction until the board is merged accordingly
and a new value is spawned.

This function returns True if a valid step was executed for False in
case of game over or an action that does not change the state. The
caller is expected to handle these cases.