Add network
Add a copy of the tutorials' network to start with. The idea is to write
the code in a way that different networks can be tried out.

I do not know how to test this code, I probably need to gain some
experience first as I don't yet fully understand how any of this
works...