Add a way to track progress
Introduce a class that can be used to track the progress of the
training. The tracked data might not prove to be enough. However, I
refrain from also tracking all the data that happens within an
individual episode (like the state transitions). I suspect this would
be quite a lot of data given that one episode will hopefully easily
contain thousands of moves, and it's likely not necessary to track this.
For sure not to judge if the training works or not.

There is some "redundancy" in the tracked data; we only need one of
number_of_moves, score or max_tile to see if the network gets better
over the course of training. I still think all 3 are valuable to have
as max_tile is very intuitive to understand while score is the better
indicator if the training works and number_of_moves seems relevant as
this may not correlate with the score, but probably should in the end.

Epsilon (just one per episode) is tracked to get a feeling on how fast
this decays, without this it's hard to tweak the parameters I could
imagine.