Add trainer
Add an initial trainer class strongly inspired by the pytorch tutorial.
It initializes all pieces needed for the trainer and keeps them in its
state.

The optimizer is intentionally not implemented yet, and the
state_to_network_input method will most probably be dependent of the
chosen network structure, and therefore it will make sense to move this
into the network class.

In contrast to the tutorial, all game simulations are done outside the
pytorch framework. The tensors will be created for the training. This
keeps the code simpler, especially for someone like me without prior
pytorch experience.

A main is added to run the training.

The training is executed as is and findings documented. The commit also
contains the data from this run.