Optimize network
Implement network optimization method. This is heavily inspired
by the pytorch tutorial on DQL. Move the 'state_to_network_input' method
functionality into the networks 'forward' method as this will
allow exchange networks much more easily.

Add a base class for networks to simplify implementing different
networks in the future. Instead of the input size, the board size is
passed as an argument into the network. This information is more
versatile.

Implement the network synchronization. This step is also "misused" to
same some network snapshots.

In the tutorial they handle cases where the "game over" state is reached
differently from the others. As this seems reasonable, introduce the
terminal state to the MemoryFragment. If an input state is terminal,
no prediction is performed as there will not be a next action.