Change training
Based on the Medium post mentioned in README.md some changes are
introduced.

In addition, the spawning of new numbers is separated from the evolving.
I want to see if the network learns better if the random factor is not
part of the training. Currently, the spawn is still part of the trained
evolved state.