Trainer: Factor out snapshotting
Factor out saving a snapshot of the network from _update_target_network.
It does not belong there, and snapshotting does not need to be that
frequent. Some networks can get big quite fast, and this is a good way
to waste time and destroy SSDs.

It's expected that thousands of episodes are needed to every 1000
episodes is probably enough. Maybe it will make sense to make this
configurable in the future as this is probably network-dependent. If
training 1000 episodes takes a long time, a shorter interval would make
sense.