Add tutorial network training with penalties
Add a training with the tutorial network with the previously described
changes (added penalties and multiple optimizations per episode).

Add a way to load a training snapshot and continue from there as I
had some issues with my GPU (or GPU driver) once. A proper retry
mechanism needs to be added if one wants to seriously use this.
Currently, the epsilon is reset when restarting. It just starts a new
training from a pretrained network.