Add convolutional network
Add the network used in the Medium blog post mentioned in the README.md.
The model and data transformation are adapted to work with arbitrary
square game sizes.

Add training results and update the README.md with some observation and
interpretation of the data.

The trained model is not added to git as it is quite large (380 MB).